pyproject.toml
[project]
name = "ai-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.9.3",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=0.3",
    "faiss-cpu>=1.8",
    "pypdf>=4.0",
    "fpdf2>=2.7",
    "python-dotenv>=1.1.1",
]
